Move working branch slice into cms slice
