+++ packages/airview-cms/src/features/cms/index.js
@@ -3,11 +3,12 @@ export { EnableCmsButton } from "./enable-cms-button";
 export {
   cmsSlice,
   setCmsContext,
+  workingBranchSlice,
+  selectWorkingBranch,
   selectCmsEnabledStatus,
   selectCmsContext,
 } from "./cms.slice";
 export { branchCreatorSlice } from "./branch-creator";
-export { workingBranchSlice, selectWorkingBranch } from "./toolbar";
 export { metaEditorSlice } from "./meta-editor";
 export { createPullRequestSlice } from "./create-pull-request";
 export { useSetCmsContext } from "./use-set-cms-context";
