+++ packages/airview-cms/src/features/cms/toolbar/index.js
@@ -1,7 +1,2 @@
 export * from "./constants";
 export { ToolBar as default } from "./toolbar";
-export {
-  workingBranchSlice,
-  setWorkingBranch,
-  selectWorkingBranch,
-} from "./working-branch.slice";
