--- packages/airview-cms/src/features/cms/toolbar/working-branch.slice.js
@@ -1,16 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-
-export const workingBranchSlice = createSlice({
-  name: "workingBranchSlice",
-  initialState: {},
-  reducers: {
-    setWorkingBranch: (state, action) => {
-      state.workingBranch = action.payload;
-    },
-  },
-});
-
-export const { setWorkingBranch } = workingBranchSlice.actions;
-
-export const selectWorkingBranch = (state) =>
-  state.workingBranchSlice.workingBranch;
